requirements.txt
browser-cookie3==0.17.0
python-dotenv==0.21.1
requests==2.28.2
aiodns==3.0.0
aiohttp==3.8.4
asyncio==3.4.3